removed JedisUtils
